Design enhanced to prevent stalling when Arduino hangs with heartbeat high
In some circumstances the Arduino can hang with the heartbeat output
held high which left the external watchdog held in the counter reset
condition thus not counting and hence no reset signal to Arduino. Design
changed to incorporate a one shot counter reset which fires when the
heartbeat transitions low.